Renamed grpc client files
