+++ services/assistant/app/infra/grpc/__init__.py
@@ -1,6 +1,6 @@
 from app.infra.grpc.client_factories import XrayClientFactory, BookingClientFactory
-from app.infra.grpc.xray_client import XrayClient
-from app.infra.grpc.booking_client import BookingClient
+from app.infra.grpc.xray_client_impl import XrayClient
+from app.infra.grpc.booking_client_impl import BookingClient
 
 __all__ = [
     "BookingClientFactory",
