+++ services/assistant/app/infra/grpc/booking_client_impl.py
